Massive update from unsigned sizes and indices to signed.

+++ include/stc/algo/csort.h
@@ -99,7 +99,7 @@ static inline void c_PASTE(cqsort_, i_tag)(i_val arr[], intptr_t lo, intptr_t hi
     }
 }
 
-static inline void c_PASTE(csort_, i_tag)(i_val arr[], size_t n)
-    { c_PASTE(cqsort_, i_tag)(arr, 0, (intptr_t)n - 1); }
+static inline void c_PASTE(csort_, i_tag)(i_val arr[], intptr_t n)
+    { c_PASTE(cqsort_, i_tag)(arr, 0, n - 1); }
 
 #include <stc/priv/template.h>
