+++ misc/examples/regex_match.c
@@ -1,6 +1,7 @@
 #define i_extern
 #include <stc/cregex.h>
 #include <stc/csview.h>
+
 #define i_val float
 #include <stc/cstack.h>
 
@@ -22,7 +23,7 @@ int main()
 
         // extract and convert all numbers in str to floats
         c_FORMATCH (i, &re, str)
-            cstack_float_push(&vec, atof(i.match[0].str));
+            cstack_float_push(&vec, (float)atof(i.match[0].str));
 
         c_FOREACH (i, cstack_float, vec)
             printf("  %g\n", *i.ref);
